pyproject.toml
[tool.poetry]
name = "queue-system"
version = "0.1.0"
description = ""
authors = ["1orange <orangel33t@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pendulum = "^2.1.2"
Flask = "^2.0.2"
psycopg2 = "2.8.6"
addict = "^2.4.0"
PyYAML = "^6.0"
Flask-RESTful = "^0.3.9"
marshmallow = "^3.14.1"
Flask-Cors = "^3.0.10"
aiosql = "^3.4.1"
flask-apispec = "0.11.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^2.9.0"
coverage = "^5.1"
pytest = "^6.2.1"
isort = "^5.7.0"
black = "^20.8b1"
pylint = "^2.5.3"
autopep8 = "1.5.7"
flake8 = "^3.8.2"
click = "8.0.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
